requirements.txt
flask
flask_cors
openai
requests
openai-whisper
